instancemanager.py:
for fix_eIM: for IM2 and IM3, all important variables are fixed after the instance is created.
For IM3, various variables for BESS and FD are fixed for the first 12h because IM3 starts on day D
Special cases for socV are handdled because it is defined from 0-24 and also dependend on it's previous state, so values for 24h are initialized but only the first 12 arefixed.

postprocess.py: store reserve penalty

previous_results_loader.py:
load_im_results: IM2 and IM3 now considered. The variables that need to be fixed after the instance is created are stored in simulation_context and are added to the instance in instancemanager.py.
In the future when I have time I will cleen up the redundant code